Removed the permission check for `cpu.pressure` item when resource manager sets to 'group-v2'.
`cpu.pressure` is one of API for PSI, but PSI(Pressure Stall Information) is an optional
feature of the Linux kernel. PSI is not enabled by default on some Linux distributions.
This causes GPDB to FAIL to start if `gp_resource_manager='group-v2'`. On the other hand,
`cpu.pressure` is not used in GPDB resource management, so it can be removed.